requirements/dev.txt
-r prod.txt
aioresponses==0.7.2
astroid==2.7.3
attrs==21.2.0
colorama==0.4.4
coverage==5.5
future==0.18.2
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
mando==0.6.4
mccabe==0.6.1
packaging==21.0
pip-licenses==3.5.2
platformdirs==2.3.0
pluggy==1.0.0
PTable==0.9.2
py==1.10.0
pylint==2.10.2
pyparsing==2.4.7
pytest==6.2.5
pytest-aiohttp==0.3.0
pytest-asyncio==0.15.1
pytest-cov==2.12.1
radon==5.1.0
six==1.16.0
toml==0.10.2
wrapt==1.12.1
